requirements.txt
beautifulsoup4==4.5.1
ntplib==0.3.3
python-dateutil==2.5.3
requests==2.11.1
pyperclip==1.5.27
flask==0.11
